// src/app/projects/models/project-filters.ts
export interface ProjectFilters {
  search?: string;
  wranglingState?: string;
  primaryWrangler?: string;
  wranglingPriority?: string;
  hasOfficialHcaPublication?: string;
  identifyingOrganism?: string;
  organOntology?: string;
  minCellCount?: string;
  maxCellCount?: string;
  searchType?: string;
}